Generalize buffers
