+++ fabric/src/main/java/com/noxcrew/noxesium/NoxesiumMod.java
@@ -11,8 +11,10 @@ import com.noxcrew.noxesium.network.serverbound.ServerboundClientSettingsPacket;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.C2SPlayChannelEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
+import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.resources.ReloadableResourceManager;
 
 import java.util.HashSet;
@@ -110,9 +112,7 @@ public class NoxesiumMod implements ClientModInitializer {
         modules.forEach(NoxesiumModule::onStartup);
 
         // Register the resource listener
-        if (Minecraft.getInstance().getResourceManager() instanceof ReloadableResourceManager reloadableResourceManager) {
-            reloadableResourceManager.registerReloadListener(new NoxesiumReloadListener());
-        }
+        ResourceManagerHelper.get(PackType.CLIENT_RESOURCES).registerReloadListener(new NoxesiumReloadListener());
     }
 
     /**
