+++ fabric/src/main/java/com/noxcrew/noxesium/feature/render/NoxesiumReloadListener.java
@@ -1,7 +1,13 @@
 package com.noxcrew.noxesium.feature.render;
 
+import com.noxcrew.noxesium.feature.render.cache.ScoreboardCache;
+import com.noxcrew.noxesium.feature.render.cache.ScoreboardInformation;
+import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
+import net.fabricmc.fabric.api.resource.SimpleSynchronousResourceReloadListener;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.PreparableReloadListener;
 import net.minecraft.server.packs.resources.ResourceManager;
+import net.minecraft.util.Unit;
 import net.minecraft.util.profiling.ProfilerFiller;
 
 import java.util.concurrent.CompletableFuture;
@@ -11,16 +17,20 @@ import java.util.concurrent.Executor;
  * Listens to Minecraft reloading the resources and clears cached scoreboard information as the
  * contents of the resource pack may have changed.
  */
-public class NoxesiumReloadListener implements PreparableReloadListener {
+public class NoxesiumReloadListener implements SimpleSynchronousResourceReloadListener {
 
     @Override
-    public CompletableFuture<Void> reload(PreparationBarrier preparationBarrier, ResourceManager resourceManager, ProfilerFiller profilerFiller, ProfilerFiller profilerFiller2, Executor executor, Executor executor2) {
-        CachedScoreboardContents.clearCache();
-        return CompletableFuture.completedFuture(null);
+    public void onResourceManagerReload(ResourceManager resourceManager) {
+        ScoreboardCache.getInstance().clearCache();
     }
 
     @Override
     public String getName() {
         return "noxesium";
+    }
+
+    @Override
+    public ResourceLocation getFabricId() {
+        return new ResourceLocation("noxesium", "reload");
     }
 }
