+++ fabric/src/main/java/com/noxcrew/noxesium/feature/render/cache/ElementInformation.java
@@ -1,0 +1,7 @@
+package com.noxcrew.noxesium.feature.render.cache;
+
+/**
+ * The basis for a record that stores information relevant to an element.
+ */
+public interface ElementInformation {
+}
