+++ fabric/src/main/java/com/noxcrew/noxesium/feature/rule/impl/DisableScoreboardNumbersRule.java
@@ -1,10 +1,7 @@
 package com.noxcrew.noxesium.feature.rule.impl;
 
-import com.noxcrew.noxesium.feature.render.CachedScoreboardContents;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screens.SoundOptionsScreen;
-
-import java.util.Objects;
+import com.noxcrew.noxesium.feature.render.cache.ScoreboardCache;
+import com.noxcrew.noxesium.feature.render.cache.ScoreboardInformation;
 
 /**
  * Updates the cached scoreboard whenever it changes.
@@ -17,6 +14,6 @@ public class DisableScoreboardNumbersRule extends BooleanServerRule {
 
     @Override
     protected void onValueChanged(Boolean oldValue, Boolean newValue) {
-        CachedScoreboardContents.clearCache();
+        ScoreboardCache.getInstance().clearCache();
     }
 }
