+++ fabric/src/main/resources/noxesium.accesswidener
@@ -43,3 +43,7 @@ accessible field net/minecraft/client/renderer/RenderType$CompositeState texture
 accessible class net/minecraft/client/renderer/RenderStateShard$TextureStateShard
 accessible field net/minecraft/client/renderer/RenderStateShard$TextureStateShard texture Ljava/util/Optional;
 accessible field net/minecraft/client/renderer/RenderStateShard name Ljava/lang/String;
+
+accessible field com/mojang/blaze3d/platform/GlStateManager BLEND Lcom/mojang/blaze3d/platform/GlStateManager$BlendState;
+accessible class com/mojang/blaze3d/platform/GlStateManager$BlendState
+accessible field com/mojang/blaze3d/platform/GlStateManager$BooleanState enabled Z
